Uneal style update in Zotero App
I made some adjustments in the style of Uneal, to meet the requirements of ABNT 6023:2018. I created two macros to insert amount of pages of theses and dissertations and pages of books, in addition to inserting the name of the thesis or dissertation advisor. This is required by ABNT, but was not being generated by Zotero. In fact, I haven't seen any style that meets those requirements yet.